requirements.txt
numpy
matplotlib
opencv-python
mediapipe
Pillow